# Conflicts:
#	src/i18n/en.json
#	src/i18n/ru.json

+++ src/routes/Players/routes/Profile/routes/Rewards/routes/FreeSpins/components/ManualFreeSpin/FreeSpinCreateModal/validator.js
@@ -22,7 +22,7 @@ export default (values, {
     },
     pageCode: ['required', 'string'],
     betLevel: ['required', 'integer'],
-    coinSize: ['required', 'integer'],
+    coinSize: ['required', 'numeric'],
     rhfpBet: ['required', 'integer'],
     comment: ['string'],
     clientId: ['string', 'required'],
